requirements.txt
Flask
requests
discord.py
python-dotenv